Fix #7 and #8: use decimal for monetary types; remove duplicate OnPropertyChanged
- Changed InterestRate and Income fields from double to decimal across all layers
  (DTOs, domain models, WPF models, service interface and implementation)
- Removed duplicate OnPropertyChanged from BaseViewModel; NotifyPropertyChanged
  (with CallerMemberName) is the single notification path used throughout

+++ Application/Dtos/DepositDto.cs
@@ -10,7 +10,7 @@ public class DepositDto
 
     public int Term { get; set; }
 
-    public double Income { get; set; }
+    public decimal Income { get; set; }
 
     public Currencies Currency { get; set; }
